BuildTools: Add make-resource-header, make-resource-source
Instead of generating the bytes as hex values in an array,
we now generate a string. We also split generation of the
header and source file, which makes it so that we don't need
to rebuild the whole bundle whenever one of the files changes.

This yields a 1.9x speedup for fresh builds (from 19.52s, to 10.36s).