AU: Create first draft of AudioManager
The purpose of this component is to ensure that we don't allocate memory
dynamically for audio. It is structured a bit like the GL renderer and
supposed to work approximately the same, i.e. if the audio is not ready,
we will provide the caller with a value of zero. This way we can keep
the audio thread does not need to block while audio is streaming in.