Ty2: Add MemoryPoker
This API fixes a buffer underflow which would sometimes occur when
sending messages from the audio thread to the io thread, since the
pages of the mailbox weren't paged in when the audio thread tried
to submit a message, causing the thread to block.

By poking at the pages periodically from a sleeper thread, we can rest
assured that when the audio thread comes along and tries to read the
memory, it will already be paged in.