Library: Rework Actor to not store the actor state
Upon using this API, I realized that most (all) of the time we were only
using one actor function and the state was mostly (all of the time)
managed by the application itself.

This patch simplifies Actor to not manage the actor state and removes
the multiple use of multiple actor functions to just one.