Ty2: Slow down memory poker 500ms => 1000ms
Memory poker is needed to ensure that the memory used by the audio
thread is paged in when accessing it, but when using the application it
seems 1000ms is just fine. This decreases the CPU load caused by the use
of multiple memory pokers which may poke at very large ranges.